Bring libvirt domain down prior to any change
When Redfish emulator is running against libvirt virtualization
backend, any changes being done to domain are not applied for
as long as the domain is up. This leads to two nuisances:

- REST API is not really REST-ful meaning that successfully
  applied change is not reflected in the document tree
- Multiple changes done to live domain XML tree may override one
  another because N-1 change done to a domain is not visible
  to N's change

The solution offered by this patch has been discovered in the
early days of computers and has been used with great success ever
after - we will simply turn the domain off and on again for the
time a change is being done to the domain.

Change-Id: I7b9d4ca23d5eb42b791c9399dbe03f921664569e